i2c: add beaglebone black backend support and cmdline options
Adds support to access an i2c adapter on the beaglebone black. It's
actually reasonably generic as it uses i2c-dev and a cmdline arg
allows selection of the i2c adapter that is accessed by gbsim.

Also add some cmdline arguments to control activation of the beaglebone
backend, i2c adapter to be used, hotplug base directory, and verbosity.

Signed-off-by: Matt Porter <mporter@linaro.org>