cports: add support for tracking cport<->protocol and exec a handler
Previously, we just called the i2c subdev handler regardless of what
traffic was arriving on the cport out endpoint. In order to support
arbitrary configurations of virtual modules and subdev handlers, we
now parse the manifest for cports and keep track of what cports and
protocols are supported on a module. Using this information, we
dispatch the appropriate subdev handler for a protocol used on a cport
when we inspect the incoming cport targetted packet from the AP.

Signed-off-by: Matt Porter <mporter@linaro.org>