gpio: add BeagleBone Black gpio backend
Add support for presenting a greybus gpio controller with 6
lines to the host. The first four gpio lines are connected
to the USR3..0 LEDs on the board. The second two gpio lines
are on the expansion connectors at P9-12 and P8-26. The LEDs
are useful for a quick check of gpio output and the other
pins can be used as inputs with a switch or a loopback to
drive and read an input for a hands-off test.

Signed-off-by: Matt Porter <mporter@linaro.org>