gbsim: fix byte order problems
All over-the-wire fields are required to be in little-endian format.
Most of our fields are one byte anyway, so they don't matter.
We're also working with little-endian architectures, so any
endianness problems in the code don't exhibit bad behavior.

But the code here is wrong, and this patch fixes all of the byte
order problems I could find.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Matt Porter <mporter@linaro.org>