gbsim: match connection cport ids
Until we establish better protocols for setting up connections, we
have been assuming certain CPort ids are in use for the AP.

For the December demo we decided that a simple way to establish this
is to have the AP side of a connection use the same CPort id as the
other end of the connection (whose CPort id is defined in the module
manifest).

Since we can no longer assume CPort id 0 in the messages we send we
need to update the code to send the right one.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Matt Porter <mporter@linaro.org>