gpio: clean up IRQ request hack
It's still a hack.  But define the size of the header in terms of
the payload data structure.  Also mark more clearly that the
operation id being sent and the GPIO number carrying the IRQ are
completly fabricated for testing purposes.

These routines would also be more useful if they dumped in pretty
form the content of their payloads.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>