main: clean up option parsing a bit
Pass a leading ':' character in the getopt option string.  This
causes getopt to distinguish between unrecognized options and
options that are not supplied their required arguments.

There was no handling of a missing hotplug_basedir argument to "-h";
add that.

Simplify the error messages a bit, and shorten some long lines.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>