gbsim: rename cport symbol names
I found the naming convention "cport_in" and "cport_out" to identify
the data that's being sent and received (respectively) by gbsim to
be confusing.  The idea was that they're named from the perspective
of the AP.  But for me it seemed to take more mental energy that it
should to keep from getting things wrong.

Rename variables and fields named "cport_in" to be "to_ap", and
"cport_out" to be "from_ap" in an effort to avoid confusion.

Change the type of the "size" variable that holds what's returned by
read() to ssize_t, so we can actually detect a negative return
value.  If an error does occur, print the error code.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>