cport: further abstract cport receive handler
Determine whether the CPort ID encoded in an incoming message is
even associated with a valid CPort before attempting to handle
it.  Pass the CPort structure rather than its id to the handler
function.

This is in preparation for creating a protocol abstraction with a
handler function (which will eliminiate cport_recv_handler()).

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>