gbsim: pass response buffer to handler
Every receive handler allocates a 4KB response buffer, fills it in,
sends it, and then frees it before returning.

Instead, have the caller pass in a 4KB buffer for the handler
function to use for sending its response.  Like the request buffer,
we can define this buffer statically and reuse it for sending every
response.

Zero the response buffer on every iteration for good measure.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>