gpio: separate interrupt testing hack
It's a little tricky to test generating an interrupt with GBSim.
Matt put in place a simple hack that sends an IRQ request when
IRQs first get unmasked.

Move that bit of code outside the switch statement.  This allows
this one case to fit the pattern used by all other cases.

In addition, the response to the GPIO IRQ operation is currently
handled in this function.  A debug message mentions its arrival, but
nothing else occurs.  Nothing else happens after this, so just
handle this other unusual case by returning directly.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>