gbsim: factor out response header creation
Every message handler begins its processing of each of its request
types by populating the response header with the size, id, type, and
result of the operation.  In almost all cases, the response is success.

We are already calculating the size using a local variable "sz" for
each of these, and the operation id and response type can be done
generically.  Assume a successful result, but allow each operation
to override that if desired.

So factor all that activity out of the handling each operation type,
and do it just once at the end of each protocol handler.  Do the
special stash of the CPort ID at the same time.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>