gbsim: consolidate accounting for response header
Every operation response message accounts for the size of its
header.  Do that at the end of handling requests in all handlers, so
the handler code can concentrate only on the message payload.
Rename the "sz" variable to be "payload_size" to emphasize this.

In the GPIO IRQ hack, dump the content of the request message before
sending it (like we do for all the responses).

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>