gbsim: convert to use autotools
I've had more errors due to .h files changed and the .c files not being
rebuilt, and issues pointing to where custom locations of libusgb and
libsoc are at, so just move to use autotools which solves all of these
issues.

The same GBDIR= "feature" is here, to point to the kernel greybus code
that is shared by the project, so no one's muscle memory should be
affected at all.

Bonus is now we can see build warnings much better, as they are obvious,
and not hidden by long gcc comand line invocations.  Patches to fix
those warnings will be forthcoming.

After this is applied, run './autogen.sh' to create the needed
configure file and then do the normal ./configure && make two-step
process we are all used to.

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>