gbsim: update the README.md file
The gbsim README file doesn't reflect the fact that we now use
autotools.

Add a section at the top to give a quick overview of what's
required to build the code.

Update the description of how GBSIM needs to be set (in the
environment, not in the Makefile).  Also update the instructions
for cross-compilation.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>