gbsim: Use cport->hd_cport_id instead of cport->id
There are two cport-ids for every cport, one is the module side id (used
to identify the cport for a interface) and the other one is AP side id
(used to identify a cport for an AP).

Currently, gbsim is using the first one to identify the second one. Its
all working because the cports are present in increasing order in
manifest and cport->id and cport->hd_cport_id are same. But as soon as
we change the sequence, this breaks.

Also keep track of hd_cport_id and use it for responses. Keep using
normal cport->id for all the print messages though.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>