gbsim: control protocol node isn't compulsory in manifest
Control protocol's node is optional in manifest and if it is missing,
things wouldn't work with current code.

The problem is allocation of hd_cport_id (i.e. AP's cport number).
gbsim allocates these ids, starting from 0, based on the order in which
cports are present in manifest, and the kernel side allocates the first
available cport id for control protocol for every interface. When the
control cport is missing in manifest, gbsim wouldn't allocate a
hd_cport_id for control protocol, but kernel side will. And this will
end up in a mismatch of ids.

Fix this by allocating a cport for control cport in gbsim, if manifest
doesn't contain it.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>