gbsim/uart: Add support for UART error signals
This patch adds support to gbsim for break, parity, framing and over-run
reporting. Recent changes to the greybus-spec move these bits to the
receive-data message. This patch adds a polling routine to gbsim such that
when parity checking is enabled gbsim will parse for break, parity and
framing errors and transmit those errors as bits in the flags field of the
receive-data message. The kernel side will then pick up those bits and
associate with TTY_BREAK, TTY_PARITY, TTY_FRAMING or TTY_OVERFLOW.

Note on beagle-bone black when the OMAP UART receives a parity or framing
error it simply inserts a NUL character into the TTY layer, so on
beaglebone black parity/framing errors are reported to gbsim in the same
way as break.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>