gbsim/uart: make receive-data and serial-state unidirectional
UART receive-data and serial-state messages have no need to receive an ACK
from the AP given the underlying Unipro network is responsible for assuring
data transmission to the AP.

This patch along with a parallel patch to the greybus specification removes
ACK wait/processing and sets the transmitted signals to unidirectional in
the id field of operations request.

This patch should be applied in conjunction with the greybus specification
update that accompanies it.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>