sdio: add sdio gbsim implementation
This implementation adds SDIO protocol support to gbsim.  It enables
testing of the SDIO greybus protocol by emulating a basic Secure
Digital card.  It does not support the entire SDIO specification.

It tests all the operations defined in the greybus specification for
the SDIO protocol, and permits use of the mmc_test driver facility to
run a set of tests over the SDIO greybus protocol. For more details
see file under test/sdio/README.md

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>
Reviewed-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>