gbsim: free cports when a module is hotplugged out
This fixes a long standing bug, where we never free the cports, created
when a module is hoplugged by copying the .mnfb file to the
hotplug-module directory.

Even if we remove the old module and copy a new one, gbsim will keep the
earlier protocol alive. And if the cport ids of both old/new module are
same, the message that are supposed to reach the protocol driver of new
module reaches the protocol driver of old module.

This does it by freeing all the cports on the response to hot-unplug
event.

This also resets hd_cport_id_count to 0, to match the id's allocated by
the AP.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>