gbsim/loopback: make test app tolerate a test that finishes quickly
Conceivably a test could finish before user-space has a chance to get into
it's inotify/select() loop and therefore user-space can miss the change on
iteration_count.

This patch reworks the test code to allow the test to finish before the
user-space application is ready to receive inotify changes. Small data sets
for example a test with of just one message drives this behaviour change.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>