inotify: properly byte swap manifest size
In get_manifest_blob(), the file size is read from the beginning of
the manifest to determine how many bytes the complete manifest
requires.  The size is in little endian format though, and we're not
properly converting the value to host byte order before using it.
Fix that.

Also, the size is two bytes, and must account for itself.  So if
it's less than 2, return failure.

Finally, lseek() can return an error, so check for that and report
it if it ever occurs.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>