inotify: don't allocate max size buffer for manifest
Currently we allocate a 64KB buffer to hold the manifest regardless
of how much is required.  We're already reading in the size of the
manifest, so by deferring allocating the buffer we can just allocate
what's needed.  Currently this buffer is leaked; that is still the
case after this patch (but a fix may be coming).

Change the type of the local variable "n" to match the type returned
by read().

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>