connection: rename cport to be connection
The "cport" structure represents what is really a "connection," and
it's a little misleading.

For example, when we allocate a "cport" we allocate a CPort ID to
represent what the AP's side of the connection is using, but we
don't actually know what the AP will actually use.  When requests
come in, we send the response to whatever CPort ID was in the
request message.  But for messages we initiate the value we are
doesn't likely match what the AP is using.  (Maybe we're not
initiating any requests?  I honestly not sure how we're sending them
to the proper CPort.)

Anyway, let's start fixing this.  Rename the "cport" structure to be
"connection".  And rename the "id" field "cport_id" (it's not a
connection id).  There are a lot of other fields that will be
renamed based on this, but we'll do this in a few steps.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>