loopback_test: make dumping of raw latency values optional
Dumping the full list of raw latency timings is probably more data than
people typically want to see except under specific circumstances. This
patch adds a "-r" switch to the command line utility to switch on dumping
of that data, in the expectation that few if any people are really looking
at that data right now anyway.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>