power_supply: add power_supply implementation
Add simulation for power_supply. This will trigger the setup of 2
batteries and 2 chargers. The properties exported by each are based in
the real properties in the hw simulated.
A proper set of values, charge/discharge simulation and online status
toggle for charges will be add afterwards to better simulate the all
power supply cycle.

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>