spi: add spi simulator
Add support for SPI master simulation. Also depending on the number of
chipselect at compile time it will create spidev for even chipselect and
spinor for odd chipselect.

The spinor is really incomplete, only the jedec (READID) operation is
implemented. For spidev it write to a file in /tmp/ and when the
operation is a write/read it sends the complement of the write bytes in
the transfer response.

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>