loopback_test: remove postfix, dev and con
Aggregation is no longer done in the kernel driver and all devices have
their own specific control files. So today, all loopback devices have the
same number and type of sysfs attributes. It is now up to the application
to start the test for each device and control which devices to use for
a particular test.

So remove the postfix, dev and con variables that were used to handle
different types of attributes depending on the device.

Reviewed-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Axel Haslam <ahaslam@baylibre.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>