loopback_test: add loopback test struct
The kernel driver no longer manages the device mask and no longer
does aggregation. The application needs to configure and set each
device it wants to exercise for a test. Having, for example, a global
control interface does not make sense anymore.

In preparation for the next couple of patches that will add a
per-device structure, remove global variables and add them to a
structure instead that is easier to pass as argument to functions.

Reviewed-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Axel Haslam <ahaslam@baylibre.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>