loopback_test: only read inotify fd when select signals data available
loopback_test was failing to detect completion of loopback tests with a
large number of iterations, e.g. about 980 iterations or greater for a
loopback transfer test with a 100 byte payload.

I traced the failure to an error in loopback_test's completion detection
logic: even when select() on an inotify fd timed out, loopback_test
nonetheless attempted to read that fd, blocking the application.

This patch changes the detection logic to only read the inotify fd (thus
clearing the notification) when select() indicates that a notification has
been posted.

Signed-off-by: Mitchell Tasman <tasman@leaflabs.com>
Reviewed-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>