gpio: add test case for irq_event
This add a new test scenario for the irq mask event and mask handling.
If a even gpio is exported and it's direction is input and with a edge
type, the next odd gpio will be connected to it (if it is configured as
output) and setting the value in the odd gpio will trigger an event to
be send to greybus driver to the even gpio.
This try to mimic a hw test case.

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>