interface: add support for interfaces and svc
This patch is long, but to avoid a break in compilation and in
functionality there is the need to make this changes in one go.

The goal here is to introduce svc and interface objects and attach
then together, jointly with the connections for each interface.
No new feature is introduced by this patch, besides moving the
point of managing the connections lifetime (creation/deletion).

This will allow for later work on supporting more than one interface
and probably more than one svc.

Signed-off-by: Rui Miguel Silva <rmfrfs@gmail.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>